+++ src/main/java/shop/donutmarket/donut/domain/myPage/service/MyPageService.java
@@ -1,6 +1,7 @@
 package shop.donutmarket.donut.domain.myPage.service;
 
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import shop.donutmarket.donut.domain.blacklist.repository.BlackListRepository;
@@ -8,9 +9,13 @@ import shop.donutmarket.donut.domain.board.repository.BoardRepository;
 import shop.donutmarket.donut.domain.myPage.dto.MyPageResp;
 import shop.donutmarket.donut.domain.payment.repository.PaymentRepository;
 import shop.donutmarket.donut.domain.report.repository.ReportRepository;
+import shop.donutmarket.donut.domain.review.model.Review;
+import shop.donutmarket.donut.domain.review.repository.ReviewRepository;
+import shop.donutmarket.donut.domain.user.model.User;
 import shop.donutmarket.donut.global.auth.MyUserDetails;
 import shop.donutmarket.donut.global.exception.Exception500;
 
+import java.util.ArrayList;
 import java.util.List;
 
 @Service
@@ -24,6 +29,8 @@ public class MyPageService {
     private final BlackListRepository blackListRepository;
 
     private final ReportRepository reportRepository;
+
+    private final ReviewRepository reviewRepository;
 
     @Transactional(readOnly = true)
     public List<MyPageResp.MyBoardDTO> 나의게시글보기(Long id) {
@@ -62,6 +69,27 @@ public class MyPageService {
             return myReportDTOS;
         } catch (Exception e) {
             throw new Exception500("나의 신고내역 보기 실패 : " + e.getMessage());
+        }
+    }
+
+    @Transactional(readOnly = true)
+    public MyPageResp.MyReviewDTO 나의리뷰목록보기(@AuthenticationPrincipal MyUserDetails myUserDetails) {
+        try {
+            List<Review> myReviews = reviewRepository.findAllByReviewerId(myUserDetails.getUser().getId());
+
+            // user 객체 만들어서 주입
+            List<Review> reviewList = new ArrayList<>();
+            for (Review reviewPS : myReviews) {
+                User reviewedUser = User.builder().name(reviewPS.getReviewed().getName()).build();
+                Review review = Review.builder().id(reviewPS.getId()).reviewed(reviewedUser).comment(reviewPS.getComment())
+                        .score(reviewPS.getScore()).createdAt(reviewPS.getCreatedAt()).build();
+                reviewList.add(review);
+            }
+            
+            MyPageResp.MyReviewDTO resp = new MyPageResp.MyReviewDTO(reviewList);
+            return resp;
+        } catch (Exception e) {
+            throw new Exception500("나의 리뷰목록 보기 실패 : " + e.getMessage());
         }
     }
 }
