+++ src/main/java/shop/donutmarket/donut/domain/review/dto/ReviewResp.java
@@ -3,6 +3,7 @@ package shop.donutmarket.donut.domain.review.dto;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.Setter;
+import shop.donutmarket.donut.domain.review.model.Review;
 
 public class ReviewResp {
