+++ src/main/java/shop/donutmarket/donut/domain/review/service/ReviewService.java
@@ -18,44 +18,34 @@ import shop.donutmarket.donut.domain.review.repository.RateRepository;
 import shop.donutmarket.donut.domain.review.repository.ReviewRepository;
 import shop.donutmarket.donut.domain.user.model.User;
 import shop.donutmarket.donut.global.auth.MyUserDetails;
+import shop.donutmarket.donut.global.exception.Exception500;
 
 @Service
 @RequiredArgsConstructor
 public class ReviewService {
-    
+
     private final ReviewRepository reviewRepository;
     private final RateRepository rateRepository;
 
-    public List<Review> 내리뷰목록(@AuthenticationPrincipal MyUserDetails myUserDetails) {
-        List<Review> myReviews = reviewRepository.findAllByReviewerId(myUserDetails.getUser().getId());
-
-        List<Review> reviewList = new ArrayList<>();
-        for (Review reviewPS : myReviews) {
-            User reviewedUser = User.builder().name(reviewPS.getReviewed().getName()).build();
-            Review review = Review.builder().id(reviewPS.getId()).reviewed(reviewedUser).comment(reviewPS.getComment())
-            .score(reviewPS.getScore()).createdAt(reviewPS.getCreatedAt()).build();
-            reviewList.add(review);
-        }
-        return reviewList;
-    }
-
     @Transactional
     public ReviewSaveRespDTO 리뷰작성(ReviewSaveReqDTO reviewSaveReqDTO, @AuthenticationPrincipal MyUserDetails myUserDetails) {
+        try {
+            // 리뷰 생성
+            Review review = Review.builder().reviewer(myUserDetails.getUser()).reviewed(reviewSaveReqDTO.getReviewedUser())
+                    .score(reviewSaveReqDTO.getScore()).comment(reviewSaveReqDTO.getComment()).createdAt(LocalDateTime.now()).build();
 
-        // 리뷰 생성
-        Review review = Review.builder().reviewer(myUserDetails.getUser()).reviewed(reviewSaveReqDTO.getReviewedUser())
-        .score(reviewSaveReqDTO.getScore()).comment(reviewSaveReqDTO.getComment()).createdAt(LocalDateTime.now()).build();
+            Review reviewPS = reviewRepository.save(review);
 
-        Review reviewPS = reviewRepository.save(review);
+            // 평점 변경 1점 추가
+            Optional<Rate> reviewedUserRateOP = rateRepository.findByUserId(reviewSaveReqDTO.getReviewedUser().getId());
 
-        // 평점 변경 1점 추가
-        Optional<Rate> reviewedUserRateOP = rateRepository.findByUserId(reviewSaveReqDTO.getReviewedUser().getId());
+            Rate rate = reviewedUserRateOP.get();
+            rate.rateUp();
 
-        Rate rate = reviewedUserRateOP.get();
-        rate.rateUp();
-
-        ReviewSaveRespDTO saveRespDTO = new ReviewSaveRespDTO(reviewPS.getScore(), reviewPS.getComment());
-        return saveRespDTO;
+            ReviewSaveRespDTO saveRespDTO = new ReviewSaveRespDTO(reviewPS.getScore(), reviewPS.getComment());
+            return saveRespDTO;
+        } catch (Exception e) {
+            throw new Exception500("내 리뷰 목록보기 실패 : " + e.getMessage());
+        }
     }
-
 }
